+++ demo/src/main/java/com/example/demo/exception/SameEmailException.java
@@ -1,7 +1,7 @@
 package com.example.demo.exception;
 
 public class SameEmailException extends  RuntimeException{
-    public SameEmailException(int id, String email) {
+    public SameEmailException(Long id, String email) {
         super("User with id : " + id + " already has this email : " + email);
     }
 }
