+++ demo/src/main/java/com/example/demo/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.demo.exception;
 
 public class UserNotFoundException extends RuntimeException{
-    public UserNotFoundException(int id) {
+    public UserNotFoundException(Long id) {
         super("User with id: " + id + " not found");
     }
 }
